fix: cast stat.Dev/Ino to uint64 for darwin cross-compilation
unix.Stat_t.Dev is int32 on Darwin but uint64 on Linux. Without an
explicit cast the struct literal fails when cross-compiling for macOS
from a Linux host. Ino is also cast for consistency since its width
varies by platform.

+++ internal/fsys/read_regular_unix.go
@@ -47,7 +47,7 @@ func (OSFS) readRegularFileSnapshot(name string) (regularFileSnapshot, error) {
 	}
 	return regularFileSnapshot{
 		data:  data,
-		id:    fileIdentity{dev: stat.Dev, ino: stat.Ino},
+		id:    fileIdentity{dev: uint64(stat.Dev), ino: uint64(stat.Ino)},
 		hasID: true,
 	}, nil
 }
